Resolves #109
