+++ .gitignore
@@ -1,5 +1,6 @@
 __pycache__/
 .vscode
+.idea
 *.py[cod]
 env/*
 .venv/*
