+++ src/models/misc/__init__.py
@@ -1,3 +1,4 @@
 from .guild import *  # noqa: F401, F403
 from .alerts import *
 from .premium import *
+from .user import *
