+++ src/models/misc/alerts.py
@@ -1,7 +1,7 @@
 from __future__ import annotations
 
 from models import BaseDbModel
-from tortoise import fields
+from tortoise import fields , models
 from models.helpers import *
 
 
@@ -35,3 +35,19 @@ class Read(BaseDbModel):
     id = fields.IntField(pk=True)
     user_id = fields.BigIntField()
     read_at = fields.DatetimeField(auto_now=True)
+
+
+class Timer(models.Model):
+    id = fields.BigIntField(pk=True)
+    expires = fields.DatetimeField(index=True)
+    created = fields.DatetimeField(auto_now=True)
+    event = fields.TextField()
+    extra = fields.JSONField(default=dict)
+
+    @property
+    def kwargs(self):
+        return self.extra.get("kwargs", {})
+
+    @property
+    def args(self):
+        return self.extra.get("args", ())
